+++ commitlint.config.js
@@ -1,10 +1,3 @@
 module.exports = {
   extends: ['@commitlint/config-conventional'],
-  rules: {
-    'subject-case': [
-      2,
-      'never',
-      ['sentence-case', 'start-case', 'pascal-case', 'upper-case', 'lower-case'],
-    ],
-  },
 };