+++ packages/walletkit/test/App.tsx
@@ -1,3 +1,4 @@
+import { useState } from 'react';
 import { chains } from './chains';
 import { WagmiConfig, createClient } from 'wagmi';
 import {
@@ -10,7 +11,6 @@ import {
   useModal,
 } from '../src/index';
 
-import { useState } from 'react';
 import {
   binanceWeb3Wallet,
   coinbaseWallet,
@@ -20,43 +20,30 @@ import {
   tokenPocket,
   trustWallet,
   walletConnect,
-  particleWallet,
-} from '../src/wallets';
-import React from 'react';
-import VConsole from 'vconsole';
-import { ParticleNetwork } from '@particle-network/auth';
+  // particleWallet,
+} from '@/wallets';
+// import { ParticleNetwork } from '@particle-network/auth';
 
-new VConsole();
+// const env = (import.meta as any).env;
 
-//!!! environment variables for testing, use directly
-const PARTICLE_APP_APP_ID = '9f8f0969-f7b3-474b-ae93-8773231e6c05';
-const PARTICLE_APP_PROJECT_ID = '33eea7b2-d76b-4b5a-978f-4413a6b70e82';
-const PARTICLE_APP_CLIENT_KEY = 'clprc7kown00uAKQrWsMOAwzXXiWxYDMq9bpfTta';
-const WALLET_CONNECT_PROJECT_ID = 'e68a1816d39726c2afabf05661a32767';
+// const particle = new ParticleNetwork({
+//   projectId: env.VITE_PARTICLE_APP_PROJECT_ID,
+//   clientKey: env.VITE_PARTICLE_APP_CLIENT_KEY,
+//   appId: env.VITE_PARTICLE_APP_APP_ID,
+//   wallet: { displayWalletEntry: true },
+//   chainId: 204,
+//   chainName: 'opBNB',
+// });
 
-const particle = new ParticleNetwork({
-  projectId: PARTICLE_APP_PROJECT_ID as string,
-  clientKey: PARTICLE_APP_CLIENT_KEY as string,
-  appId: PARTICLE_APP_APP_ID as string,
-  wallet: { displayWalletEntry: true },
-  chainId: 204,
-  chainName: 'opBNB',
-});
-
-particle.setERC4337({
-  name: 'BICONOMY',
-  version: '2.0.0',
-});
+// particle.setERC4337({
+//   name: 'BICONOMY',
+//   version: '2.0.0',
+// });
 
 const client = createClient(
   getDefaultConfig({
     autoConnect: true,
     appName: 'WalletKit',
-
-    // WalletConnect 2.0 requires a projectId which you can create quickly
-    // and easily for free over at WalletConnect Cloud https://cloud.walletconnect.com/sign-in
-    walletConnectProjectId: WALLET_CONNECT_PROJECT_ID,
-
     chains,
     connectors: [
       trustWallet(),
@@ -66,7 +53,7 @@ const client = createClient(
       binanceWeb3Wallet(),
       coinbaseWallet(),
       tokenPocket(),
-      particleWallet(),
+      // particleWallet(),
       walletConnect(),
     ],
   }),